[Rename][BE] 패키지명 소문자, 알맞게 변경
