+++ backend/src/main/java/com/ll/hotel/domain/booking/booking/controller/BookingController.java
@@ -3,8 +3,8 @@ package com.ll.hotel.domain.booking.booking.controller;
 import com.ll.hotel.domain.booking.booking.dto.*;
 import com.ll.hotel.domain.booking.booking.service.BookingService;
 import com.ll.hotel.domain.member.member.entity.Member;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.standard.page.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
