+++ backend/src/main/java/com/ll/hotel/domain/hotel/room/controller/RoomController.java
@@ -10,8 +10,8 @@ import com.ll.hotel.domain.hotel.room.dto.PutRoomResponse;
 import com.ll.hotel.domain.hotel.room.service.RoomService;
 import com.ll.hotel.domain.image.type.ImageType;
 import com.ll.hotel.domain.member.member.entity.Member;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
