+++ backend/src/main/java/com/ll/hotel/domain/member/admin/controller/AdminBusinessController.java
@@ -4,9 +4,7 @@ import com.ll.hotel.domain.member.admin.dto.request.AdminBusinessRequest;
 import com.ll.hotel.domain.member.admin.dto.response.AdminBusinessResponse;
 import com.ll.hotel.domain.member.admin.service.AdminBusinessService;
 import com.ll.hotel.domain.member.member.entity.Business;
-import com.ll.hotel.global.exceptions.ErrorCode;
-import com.ll.hotel.global.exceptions.ServiceException;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.standard.page.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
