+++ backend/src/main/java/com/ll/hotel/domain/member/admin/controller/AdminHotelController.java
@@ -4,8 +4,7 @@ import com.ll.hotel.domain.hotel.hotel.entity.Hotel;
 import com.ll.hotel.domain.member.admin.dto.request.AdminHotelRequest;
 import com.ll.hotel.domain.member.admin.dto.response.AdminHotelResponse;
 import com.ll.hotel.domain.member.admin.service.AdminHotelService;
-import com.ll.hotel.global.exceptions.ServiceException;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.standard.page.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
