+++ backend/src/main/java/com/ll/hotel/domain/member/member/controller/OAuth2Controller.java
@@ -1,8 +1,8 @@
 package com.ll.hotel.domain.member.member.controller;
 
 import com.ll.hotel.domain.member.member.entity.Member;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.global.security.oauth2.entity.OAuth;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
