+++ backend/src/main/java/com/ll/hotel/domain/member/member/service/MemberService.java
@@ -10,8 +10,8 @@ import com.ll.hotel.domain.member.member.repository.MemberRepository;
 import com.ll.hotel.domain.member.member.type.MemberStatus;
 import com.ll.hotel.global.jwt.dto.GeneratedToken;
 import com.ll.hotel.global.jwt.dto.JwtProperties;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.global.security.oauth2.entity.OAuth;
 import com.ll.hotel.global.security.oauth2.repository.OAuthRepository;
 import com.ll.hotel.standard.util.Ut;
