+++ backend/src/main/java/com/ll/hotel/domain/review/comment/controller/ReviewCommentController.java
@@ -3,7 +3,7 @@ package com.ll.hotel.domain.review.comment.controller;
 import com.ll.hotel.domain.member.member.entity.Member;
 import com.ll.hotel.domain.review.comment.dto.request.ReviewCommentContentRequest;
 import com.ll.hotel.domain.review.comment.service.ReviewCommentService;
-import com.ll.hotel.global.rq.Rq;
+import com.ll.hotel.global.request.Rq;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
