+++ backend/src/main/java/com/ll/hotel/domain/review/review/controller/ReviewController.java
@@ -9,8 +9,8 @@ import com.ll.hotel.domain.review.review.dto.response.MyReviewResponse;
 import com.ll.hotel.domain.review.review.dto.response.PresignedUrlsResponse;
 import com.ll.hotel.domain.review.review.service.ReviewService;
 import com.ll.hotel.global.app.AppConfig;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import com.ll.hotel.standard.page.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
