--- backend/src/main/java/com/ll/hotel/global/initData/BaseInit.java
@@ -1,2 +1,0 @@
-package com.ll.hotel.global.initData;
-public class BaseInit {}
