+++ backend/src/main/java/com/ll/hotel/global/initdata/BaseInit.java
@@ -1,0 +1,2 @@
+package com.ll.hotel.global.initdata;
+public class BaseInit {}
