+++ backend/src/test/java/com/ll/hotel/global/jwt/JwtAuthenticationTest.java
@@ -9,8 +9,8 @@ import com.ll.hotel.domain.member.member.type.MemberStatus;
 import com.ll.hotel.global.exceptions.ServiceException;
 import com.ll.hotel.global.jwt.dto.GeneratedToken;
 import com.ll.hotel.global.jwt.dto.JwtProperties;
-import com.ll.hotel.global.rq.Rq;
-import com.ll.hotel.global.rsData.RsData;
+import com.ll.hotel.global.request.Rq;
+import com.ll.hotel.global.response.RsData;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.Cookie;
